+++ internal/view/table.go
@@ -60,7 +60,7 @@ func (t *Table) Init(ctx context.Context) (err error) {
 	}
 	t.SetInputCapture(t.keyboard)
 	t.bindKeys()
-	t.GetModel().SetRefreshRate(time.Duration(t.app.Config.K9s.GetRefreshRate()) * time.Second)
+	t.GetModel().SetRefreshRate(time.Duration(t.app.Config.K9s.GetRefreshRate() * float64(time.Second)))
 	t.CmdBuff().AddListener(t)
 
 	return nil
